+++ NEWS.md
@@ -1,6 +1,4 @@
 
-
-
 # Ramses 0.3.0
 
 ## Features
@@ -15,6 +13,18 @@
 ## Documentation
 
 * new ramses-objects vignette
+
+# Ramses 0.2.1
+
+## Bug 
+
+* Corrected incorrect transitive closure settings
+* Corrected prescription edge classification
+
+## Improvement
+
+* Complete verification of prescription edge classification
+* therapy_id now always set to the id of the first prescription in the episode
 
 # Ramses 0.2.0